requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
PyJWT==2.9.0
